+++ lib/src/adaptive_refresh_indicator/cupertino_sliver_refresh_control_configuration.dart
@@ -1,16 +1,14 @@
-
 import "package:flutter/cupertino.dart";
 
 class CupertinoSliverRefreshControlConfiguration {
-
   final double refreshTriggerPullDistance;
 
   final double refreshIndicatorExtent;
 
   final RefreshControlIndicatorBuilder? builder;
 
-  static const double _defaultRefreshTriggerPullDistance = 100.0;
-  static const double _defaultRefreshIndicatorExtent = 60.0;
+  static const double _defaultRefreshTriggerPullDistance = 100;
+  static const double _defaultRefreshIndicatorExtent = 60;
 
   /// see [CupertinoSliverRefreshControlConfiguration] for all descriptions.
   ///
@@ -20,5 +18,4 @@ class CupertinoSliverRefreshControlConfiguration {
     this.refreshIndicatorExtent = _defaultRefreshIndicatorExtent,
     this.builder = CupertinoSliverRefreshControl.buildRefreshIndicator,
   });
-
 }