+++ lib/src/warn_before_unload/warn_before_unload.dart
@@ -1,0 +1,3 @@
+export "warn_before_unload_web.dart"
+    if (dart.library.io) "warn_before_unload_io.dart"
+    if (dart.library.html) "warn_before_unload_web.dart";